requirements.txt
PyQt6>=6.7.0
faster-whisper>=0.10.0
torch>=2.2.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
pyinstaller>=6.5.0
